+++ Example/App.js
@@ -66,11 +66,21 @@ class Expandable_ListView extends Component {
             case 3:
                 CleverTap.profileRemoveMultiValueForKey('b', 'letters');
                 break;
-            case 4: //Removing a Value from the Multiple Values
+            case 4:
                 CleverTap.profileRemoveMultiValueForKey('b', 'letters');
                 break;
             case 5:
                 CleverTap.profileAddMultiValueForKey('d', 'letters');
+                break;
+            case 500:
+                CleverTap.profileIncrementValueForKey(10, 'score');
+                CleverTap.profileIncrementValueForKey(3.141,'PI_Float');
+                CleverTap.profileIncrementValueForKey(3.141592653589793,'PI_Double');
+                break;
+            case 501:
+                CleverTap.profileDecrementValueForKey(10, 'score');
+                CleverTap.profileDecrementValueForKey(3.141, 'PI_Float');
+                CleverTap.profileDecrementValueForKey(3.141592653589793, 'PI_Double');
                 break;
             case 6:
                 onUser_Login();
@@ -201,6 +211,15 @@ class Expandable_ListView extends Component {
             case 45:
                 getFeatureFlag();
                 break;
+            case 450:
+                CleverTap.suspendInAppNotifications();
+                break;
+            case 451:
+                CleverTap.discardInAppNotifications();
+                break;
+            case 452:
+                CleverTap.resumeInAppNotifications();
+                break;
             case 46:
                 enablePersonalization();
                 break;
@@ -290,7 +309,7 @@ export default class App extends Component {
         // Listener to handle incoming deep links
         Linking.addEventListener('url', _handleOpenUrl);
 
-        // this handles the case where a deep link launches the application
+        /// this handles the case where a deep link launches the application
         Linking.getInitialURL().then((url) => {
             if (url) {
                 console.log('launch url', url);
@@ -317,7 +336,8 @@ export default class App extends Component {
                 sub_Category: [{ id: 1, name: 'pushProfile' }, { id: 2, name: 'set Multi Values For Key' }, {
                     id: 3,
                     name: 'removeMultiValueForKey'
-                }, { id: 4, name: 'removeValueForKey' }, { id: 5, name: 'addMultiValueForKey' }]
+                }, { id: 4, name: 'removeValueForKey' }, { id: 5, name: 'addMultiValueForKey' }
+                    , { id: 500, name: 'Increment Value' }, { id: 501, name: 'Decrement Value' }]
             },
 
             {
@@ -397,6 +417,10 @@ export default class App extends Component {
                 expanded: false, category_Name: "Feature Flag", sub_Category: [{ id: 45, name: 'getFeatureFlag' }]
             },
             {
+                expanded: false, category_Name: "InApp Controls", sub_Category: [{ id: 450, name: 'suspendInAppNotifications' }
+                    , { id: 451, name: 'discardInAppNotifications' }, { id: 452, name: 'resumeInAppNotifications' }]
+            },
+            {
                 expanded: false,
                 category_Name: "App Personalisation",
                 sub_Category: [{ id: 46, name: 'enablePersonalization' }, { id: 47, name: 'get profile Property' }]
@@ -404,7 +428,7 @@ export default class App extends Component {
             {
                 expanded: false,
                 category_Name: "Attributions",
-                sub_Category: [{ id: 48, name: 'get CleverTap Attribution Identifier' }]
+                sub_Category: [{ id: 48, name: '(Deprecated) get CleverTap Attribution Identifier' }]
             },
             {
                 expanded: false,
@@ -507,8 +531,14 @@ removeValueForKey = () => {
 
 };
 getCleverTap_id = () => {
+    // Below method is deprecated since 0.6.0, please check index.js for deprecation, instead use CleverTap.getCleverTapID()
+    /*CleverTap.profileGetCleverTapID((err, res) => {
+        console.log('CleverTapID', res, err);
+        alert(`CleverTapID: \n ${res}`);
+    });*/
 
-    CleverTap.profileGetCleverTapID((err, res) => {
+    // Use below newly added method
+    CleverTap.getCleverTapID((err, res) => {
         console.log('CleverTapID', res, err);
         alert(`CleverTapID: \n ${res}`);
     });
@@ -858,7 +888,7 @@ profile_getProperty = () => {
 ///Attributions
 GetCleverTapAttributionIdentifier = () => {
 
-
+    // Below method is deprecated since 0.6.0, please check index.js for deprecation, use CleverTap.getCleverTapID(callback) instead
     //Default Instance
     CleverTap.profileGetCleverTapAttributionIdentifier((err, res) => {
         console.log('CleverTapAttributionIdentifier', res, err);
@@ -1034,7 +1064,7 @@ const styles = StyleSheet.create({
     },
 
     button_Text: {
-        width:'100%',
+        width: '100%',
         textAlign: 'center',
         color: '#000',
         fontWeight: 'bold',
@@ -1042,7 +1072,7 @@ const styles = StyleSheet.create({
     },
 
     setSubCategoryFontSizeOne: {
-        fontSize: 18 
+        fontSize: 18
     },
 
 });